requirements.txt
langchain==0.3.25
langchain-google-genai==2.1.4
langgraph==0.4.1
python-dotenv==1.1.0